requirements.txt
wheel
build
twine
black
flake8
pytest
pytest-cov
setuptools
pre-commit
tensorflow
PyYAML
Pillow
tqdm
pandas